requirements.txt
requests==2.31.0
PyGithub==2.1.1
python-dotenv==1.0.0
cryptography==41.0.7
pytest==7.4.3
